+++ src/axom/mint/mesh/UnstructuredMesh.hpp
@@ -708,7 +708,7 @@ public:
   virtual IndexType getNumberOfCellNodes(IndexType cellID = 0) const final override
   {
     using CardinalityPolicy = typename CellToNodeRelation::CardinalityPolicy;
-    return m_cell_node_rel.CardinalityPolicy::size(cellID);
+    return static_cast<CardinalityPolicy>(m_cell_node_rel).size(cellID);
   }
 
   /*!
@@ -888,7 +888,7 @@ public:
   virtual IndexType getNumberOfFaceNodes(IndexType faceID = 0) const final override
   {
     using CardinalityPolicy = typename FaceToNodeRelation::CardinalityPolicy;
-    return m_face_node_rel.CardinalityPolicy::size(faceID);
+    return static_cast<CardinalityPolicy>(m_face_node_rel).size(faceID);
   }
 
   /*!