+++ src/thirdparty/tests/raja_smoke.cpp
@@ -80,6 +80,20 @@ AXOM_CUDA_TEST(raja_smoke, basic_use)
 
   axom::setDefaultAllocator(prev_allocator);
 #endif
+
+#if defined(AXOM_USE_HIP) && defined(RAJA_ENABLE_HIP) && \
+  defined(AXOM_USE_UMPIRE)
+  const int prev_allocator = axom::getDefaultAllocatorID();
+  const int UnifiedAllocatorID =
+    axom::getUmpireResourceAllocatorID(umpire::resource::Unified);
+  axom::setDefaultAllocator(UnifiedAllocatorID);
+
+  std::cout << "Testing RAJA HIP execution" << std::endl;
+  constexpr int BLOCKSIZE = 256;
+  raja_basic_usage_test<RAJA::hip_exec<BLOCKSIZE>>();
+
+  axom::setDefaultAllocator(prev_allocator);
+#endif
 }
 
 //------------------------------------------------------------------------------